update cart controller- sync get cartitem with frontend

+++ cart/src/main/java/com/lexuancong/cart/viewmodel/product/ProductPreviewVm.java
@@ -1,7 +1,9 @@
 package com.lexuancong.cart.viewmodel.product;
 
+import java.math.BigDecimal;
+
 public record ProductPreviewVm(
-        Long id,String name,String slug,Double price ,String avatarUrl) {
+        Long id, String name, String slug, BigDecimal price , String avatarUrl) {
 
 
 }
